Crate mc_storage

Storage overrides readers

In order for the client to access on pallets chain data data it has to read from the storage. This can be achieve either through the pallet runtime API or by indexing the storage directly. The OverrideHandle make it possible to use the later, more efficient way, while keeping the first one as a fallback. It can also support multiple versions of the pallet storage.

Structs

OverrideHandle

A handle containing multiple entities implementing StorageOverride

RuntimeApiStorageOverride

A wrapper type for the Runtime API.

SchemaV1Override

An override for runtimes that use Schema V1

Traits

StorageOverride

Something that can fetch Starknet-related data. This trait is quite similar to the runtime API, and indeed the implementation of it uses the runtime API. Having this trait is useful because it allows optimized implementations that fetch data from a State Backend with some assumptions about pallet-starknet’s storage schema. Using such an optimized implementation avoids spawning a runtime and the overhead associated with it.

Functions

onchain_storage_schema

Retrieve the current pallet-starknet storage schema version

overrides_handle

Create and return a handle of the starknet schema overrides