Trait mc_storage::StorageOverride

pub trait StorageOverride<B: BlockT>: Send + Sync {
    // Required methods
    fn get_storage_by_storage_key(
        &self,
        block_hash: B::Hash,
        address: ContractAddress,
        key: StorageKey
    ) -> Option<StarkFelt>;
    fn contract_class_hash_by_address(
        &self,
        block_hash: B::Hash,
        address: ContractAddress
    ) -> Option<ClassHash>;
    fn contract_class_by_address(
        &self,
        block_hash: B::Hash,
        address: ContractAddress
    ) -> Option<ContractClass>;
    fn contract_class_by_class_hash(
        &self,
        block_hash: B::Hash,
        contract_class_hash: ClassHash
    ) -> Option<ContractClass>;
    fn nonce(
        &self,
        block_hash: B::Hash,
        address: ContractAddress
    ) -> Option<Nonce>;
}

Something that can fetch Starknet-related data. This trait is quite similar to the runtime API, and indeed the implementation of it uses the runtime API. Having this trait is useful because it allows optimized implementations that fetch data from a State Backend with some assumptions about pallet-starknet’s storage schema. Using such an optimized implementation avoids spawning a runtime and the overhead associated with it.

Required Methods

fn get_storage_by_storage_key( &self, block_hash: B::Hash, address: ContractAddress, key: StorageKey ) -> Option<StarkFelt>

get storage

fn contract_class_hash_by_address( &self, block_hash: B::Hash, address: ContractAddress ) -> Option<ClassHash>

Return the class hash at the provided address for the provided block.

fn contract_class_by_address( &self, block_hash: B::Hash, address: ContractAddress ) -> Option<ContractClass>

Return the contract class at the provided address for the provided block.

fn contract_class_by_class_hash( &self, block_hash: B::Hash, contract_class_hash: ClassHash ) -> Option<ContractClass>

Return the contract class for a provided class_hash and block hash.

fn nonce(&self, block_hash: B::Hash, address: ContractAddress) -> Option<Nonce>

Returns the nonce for a provided contract address and block hash.